[Docs]: add badge dropdown example
