+++ app/src/docs/examples/badge/DropdownExample.scss
@@ -1,0 +1,9 @@
+.badgeStyles {
+    width: 107px;
+}
+
+.dot {
+    height: 12px;
+    width: 12px;
+    border-radius: 50%;
+}
