--- app/src/docs/examples/badge/TypesExample.scss
@@ -1,3 +1,0 @@
-.badgeStyles {
-    width: 107px;
-}
