+++ public/docs/content/examples-badge-Dropdown.json
@@ -1,0 +1,21 @@
+{
+  "object": "value",
+  "document": {
+    "object": "document",
+    "data": {},
+    "nodes": [
+      {
+        "object": "block",
+        "type": "paragraph",
+        "data": {},
+        "nodes": [
+          {
+            "object": "text",
+            "text": "Dropdown is a most common usecase to show status selection options",
+            "marks": []
+          }
+        ]
+      }
+    ]
+  }
+}
